Move class to new file
